+++ web/navigation/Sidebar.tsx
@@ -1,15 +1,27 @@
 import React, { memo } from 'react';
 
-import { VStack } from '../layout/VStack';
+import { gutter } from '@cbhq/cds-common/tokens/sizing';
 
-type SidebarProps = {
-  logo: React.ReactNode;
+import { VStack } from '../layout/VStack';
+import { LogoProps } from './Logo';
+import { SidebarItemProps } from './SidebarItem';
+
+export type SidebarProps = {
+  logo: React.ReactElement<LogoProps>;
+  children: React.ReactElement<SidebarItemProps>[];
 };
 
 export const Sidebar: React.FC<SidebarProps> = memo(({ logo, children }) => {
   return (
-    <VStack borderedEnd height="100%" width="100%" spacingHorizontal={2} spacingVertical={3}>
-      <VStack spacingStart={2} spacingBottom={3}>
+    <VStack
+      borderedEnd
+      height="100%"
+      width="100%"
+      spacingHorizontal={2}
+      spacingBottom={2}
+      spacingTop={gutter}
+    >
+      <VStack spacingTop={0.5} spacingStart={1} spacingBottom={3}>
         {logo}
       </VStack>
       {children}