+++ web/overlays/tooltipStyles.ts
@@ -5,7 +5,7 @@ import { palette } from '../tokens';
 
 export const container = css`
   background-color: ${palette.foreground};
-  border-radius: ${borderRadius.tooltip};
+  border-radius: ${borderRadius.tooltip}px;
   z-index: 1000;
   min-height: 24px;
   align-items: center;